speed and direction of movement figure revision
original figure 3 included movement of 10 countries on the path.  the figure has been overhauled to include a speed and direction of movement analysis and figure of all countries and regions.  the 10 country figure was revised and moved to the supplement.  I additionally revised the comments to the updated figure numbers.